refactor: LayoutConstants and localization fixes
