+++ iOS-FakeNFT-Extended/DesignSystem/LayoutConstants.swift
@@ -1,0 +1,20 @@
+//
+//  Untitled.swift
+//  iOS-FakeNFT-Extended
+//
+//  Created by Denis on 27/01/2026.
+//
+
+import SwiftUI
+
+enum LayoutConstants {
+    static let buttonHeightSmall: CGFloat = 44
+    static let buttonHeightLarge: CGFloat = 60
+    static let cornerRadius: CGFloat = 12
+    static let nftImageSize: CGFloat = 108
+    
+    static let paddingStandard: CGFloat = 16
+    static let paddingSmall: CGFloat = 8
+    static let paddingMedium: CGFloat = 12
+    static let paddingLarge: CGFloat = 20
+}
