+++ iOS-FakeNFT-Extended/Resources/en.lproj/Localizable.strings
@@ -1,28 +1,28 @@
-"Tab.catalog" = "Каталог";
-"Tab.cart" = "Корзина";
-"Catalog.openNft" = "Открыть Nft";
-"Error.network" = "Произошла ошибка сети";
-"Error.unknown" = "Произошла неизвестная ошибка";
-"Error.repeat" = "Повторить";
-"Error.title" = "Ошибка";
-"Order.empty.title" = "Корзина пуста";
-"Order.payment.button" = "К оплате";
-"Order.row.price" = "Цена";
-"Order.delete.confirmation" = "Вы уверены, что хотите удалить объект из корзины?";
-"Order.delete.confirm" = "Удалить";
-"Order.delete.cancel" = "Вернуться";
-"Order.sorting.title" = "Сортировка";
-"Order.sorting.byPrice" = "По цене";
-"Order.sorting.byRating" = "По рейтингу";
-"Order.sorting.byName" = "По названию";
-"Order.sorting.cancel" = "Закрыть";
-"CurrencySelection.title" = "Выберите способ оплаты";
-"CurrencySelection.agreementText" = "Совершая покупку, вы соглашаетесь с условиями";
-"CurrencySelection.termsOfUseLink" = "Пользовательского соглашения";
-"CurrencySelection.payButton" = "Оплатить";
-"CurrencySelection.close" = "Закрыть";
-"Payment.error.title" = "Не удалось произвести оплату";
-"Payment.error.cancel" = "Отмена";
-"Payment.error.retry" = "Повторить";
-"Payment.success.title" = "Успех! Оплата прошла, поздравляем с покупкой!";
-"Payment.success.backToCart" = "Вернуться в корзину";
+"Tab.catalog" = "Catalog";
+"Tab.cart" = "Cart";
+"Catalog.openNft" = "Open NFT";
+"Error.network" = "A network error occurred";
+"Error.unknown" = "An unknown error occurred";
+"Error.repeat" = "Retry";
+"Error.title" = "Error";
+"Order.empty.title" = "Cart is empty";
+"Order.payment.button" = "To payment";
+"Order.row.price" = "Price";
+"Order.delete.confirmation" = "Are you sure you want to remove this item from the cart?";
+"Order.delete.confirm" = "Delete";
+"Order.delete.cancel" = "Back";
+"Order.sorting.title" = "Sorting";
+"Order.sorting.byPrice" = "By price";
+"Order.sorting.byRating" = "By rating";
+"Order.sorting.byName" = "By name";
+"Order.sorting.cancel" = "Close";
+"CurrencySelection.title" = "Select payment method";
+"CurrencySelection.agreementText" = "By making a purchase, you agree to the";
+"CurrencySelection.termsOfUseLink" = "Terms of Use";
+"CurrencySelection.payButton" = "Pay";
+"CurrencySelection.close" = "Close";
+"Payment.error.title" = "Payment failed";
+"Payment.error.cancel" = "Cancel";
+"Payment.error.retry" = "Retry";
+"Payment.success.title" = "Success! Payment completed, congratulations on your purchase!";
+"Payment.success.backToCart" = "Back to cart";
