+++ iOS-FakeNFT-Extended/Resources/ru.lproj/Localizable.strings
@@ -1,0 +1,28 @@
+"Tab.catalog" = "Каталог";
+"Tab.cart" = "Корзина";
+"Catalog.openNft" = "Открыть Nft";
+"Error.network" = "Произошла ошибка сети";
+"Error.unknown" = "Произошла неизвестная ошибка";
+"Error.repeat" = "Повторить";
+"Error.title" = "Ошибка";
+"Order.empty.title" = "Корзина пуста";
+"Order.payment.button" = "К оплате";
+"Order.row.price" = "Цена";
+"Order.delete.confirmation" = "Вы уверены, что хотите удалить объект из корзины?";
+"Order.delete.confirm" = "Удалить";
+"Order.delete.cancel" = "Вернуться";
+"Order.sorting.title" = "Сортировка";
+"Order.sorting.byPrice" = "По цене";
+"Order.sorting.byRating" = "По рейтингу";
+"Order.sorting.byName" = "По названию";
+"Order.sorting.cancel" = "Закрыть";
+"CurrencySelection.title" = "Выберите способ оплаты";
+"CurrencySelection.agreementText" = "Совершая покупку, вы соглашаетесь с условиями";
+"CurrencySelection.termsOfUseLink" = "Пользовательского соглашения";
+"CurrencySelection.payButton" = "Оплатить";
+"CurrencySelection.close" = "Закрыть";
+"Payment.error.title" = "Не удалось произвести оплату";
+"Payment.error.cancel" = "Отмена";
+"Payment.error.retry" = "Повторить";
+"Payment.success.title" = "Успех! Оплата прошла, поздравляем с покупкой!";
+"Payment.success.backToCart" = "Вернуться в корзину";
